python-sop-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database dependencies
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Machine Learning and NLP
sentence-transformers==2.7.0
numpy==1.24.3
scikit-learn==1.3.2

# CrewAI and LLM dependencies
crewai==0.28.0
crewai-tools==0.1.6
langchain-openai==0.1.1
openai==1.12.0

# Document processing
python-docx==1.1.0
markdown==3.5.1

# Utility dependencies
aiofiles==23.2.1